Remove a circular dependency
Also: removed two unused types in import, and remove a useless await
keyword.
